CheckUserGetEdits: normalise IPs in 'get edits' tab
Uses IPUtils::prettifyIP() to make sure the same IPs from a technical POV are also formatted the same. Applying it for the links as well and not only the displayed texts also helps with the :visited pseudoclass
A side effect of this is that until now most IPv6 addresses were displayed in uppercase in the results, and this will be a visible change for frequent users of the tool.

Includes a test for IPv6 normalisation

Change-Id: I68e3a1403b12f4b42f2b5a2e78de5ea6d05e5149
Bug: T326996